pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "secure-auth-system"
version = "1.0.0"
description = "A secure authentication system with advanced features"
requires-python = ">=3.9"
dependencies = [
    "Django>=5.0.1",
    "django-crispy-forms>=2.1",
    "python-dotenv>=1.0.0",
    "pyotp>=2.9.0",
    "cryptography>=41.0.0",
    "django-filter>=23.5",
    "djangorestframework>=3.14.0",
    "user-agents>=2.2.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
python_files = ["test_*.py", "*_test.py"]
addopts = "-ra -q --cov=accounts"
testpaths = ["tests"]

[tool.coverage.run]
source = ["accounts"]
omit = ["*/migrations/*", "*/tests/*"]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3